Move permissionConfirm to views
