+++ app/src/router.js
@@ -1,7 +1,6 @@
 import Vue from 'vue'
 import Router from 'vue-router'
 
-import PermissionConfirm from './components/Confirm/PermissionConfirm'
 import WalletHistory from './containers/WalletHistory'
 import { WalletHome, WalletHomeCollectible, WalletHomeMain } from './containers/WalletHome'
 import WalletSettings from './containers/WalletSettings'
@@ -17,6 +16,7 @@ import WalletTransfer from './containers/WalletTransfer'
 import store from './store'
 import Confirm from './views/Confirm'
 import Login from './views/Login'
+import PermissionConfirm from './views/PermissionConfirm'
 import Popup from './views/Popup'
 import ProviderChange from './views/ProviderChange'
 import RedirectCatch from './views/RedirectCatch'
