+++ app/src/views/PermissionConfirm/PermissionConfirm.scss
