+++ app/src/views/PermissionConfirm/index.js
